Enumerators and exception moved to separate files
Also:
- using apple_in_app from plan metadata as the source of product ID
- validation in case the plan is missing the required metadata
